frame-embedder/requirements.txt
torch
transformers
Pillow
bubbaloop-sdk @ git+https://github.com/kornia/bubbaloop.git#subdirectory=python-sdk
